backend/auth-service/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-dotenv
httpx
python-multipart
pypdf
docx2txt
google-genai
geopy
